+++ BoomProject/BoomProject/ViewController/LoginViewController.m
@@ -123,7 +123,7 @@
                                              }
                                              onFailure:^(NSError *error, NSInteger statusCode) {
                                                  [self connectionErrorAlert];
-                                                 NSLog(@"error = %@, code = %ld", [error localizedDescription], statusCode);
+                                                 NSLog(@"error = %@, code = %ld", [error localizedDescription], (long)statusCode);
                                              }];
 }
